fix make adjust encounter rates

+++ include/pokemon.h
@@ -781,7 +781,7 @@ typedef struct WildEncounterWork {
     u16 minLevel;
 } WildEncounterWork;
 
-void LONG_CALL arrayShuffle(u8 array[], int n);
+void LONG_CALL arrayShuffle(int array[], int n);
 
 /**
  *  @brief allocate enough space for a PartyPokemon and zero it out
