+++ src/pokemon.c
@@ -25,11 +25,11 @@ extern u32 word_to_store_form_at;
 // [preevo] = {species, form}, [postevo] = {species, form},
 u16 ALIGN4 gEvolutionSceneOverride[2][2];
 
-void LONG_CALL arrayShuffle(u8 array[], int n)
+void LONG_CALL arrayShuffle(int array[], int n)
 {
     for (int i = n - 1; i > 0; i--) {
         int j = gf_rand() % (i + 1);
-        u8 temp = array[i];
+        int temp = array[i];
         array[i] = array[j];
         array[j] = temp;
     }
@@ -1397,7 +1397,7 @@ BOOL LONG_CALL GiveMon(int heapId, void *saveData, int species, int level, int f
 
 #ifdef RANDOM_3_MAX_IVS
     if (CheckScriptFlag(RANDOM_3_MAX_IVS_FLAG) == 1) {
-        u8 array[] = { 0, 1, 2, 3, 4, 5 };
+        int array[] = { 0, 1, 2, 3, 4, 5 };
         arrayShuffle(array, 6);
 
         int iv = 31;
@@ -1513,7 +1513,7 @@ void LONG_CALL CreateBoxMonData(struct BoxPokemon *boxmon, int species, int leve
 #ifdef RANDOM_3_MAX_IVS
         if (CheckScriptFlag(RANDOM_3_MAX_IVS_FLAG) == 1 && CheckScriptFlag(164) == 1)
         {
-            u8 array[] = { 0, 1, 2, 3, 4, 5 };
+            int array[] = { 0, 1, 2, 3, 4, 5 };
             arrayShuffle(array, 6);
 
             int iv = 31;
@@ -1608,7 +1608,7 @@ void set_starter_hidden_ability(struct Party *party UNUSED, struct PartyPokemon 
 
 #ifdef RANDOM_3_MAX_IVS
     if (CheckScriptFlag(RANDOM_3_MAX_IVS_FLAG) == 1) {
-        u8 array[] = { 0, 1, 2, 3, 4, 5 };
+        int array[] = { 0, 1, 2, 3, 4, 5 };
         arrayShuffle(array, 6);
 
         int iv = 31;
